refactor(gui): extract pure helpers from MainWindow.cpp to MainWindowHelpers (#3351 Phase 0)
First PR of the #3351 monolith-decomposition series. Pure code motion —
no behavior change.

Moves the ~20 file-scope static helper functions that have no MainWindow
dependency (no members, no mutable file-scope state) out of
MainWindow.cpp into a new MainWindowHelpers.{h,cpp}:

- Network diagnostics tooltip: formatNetworkMs, formatNetworkSeqErrors,
  buildNetworkTooltip
- TNF tooltip: tnfFrequencyHz, formatTnfFrequency, formatTnfDepth,
  buildTnfTooltip
- Memory/passive spot-ID math: memorySpotId, memoryIndexFromSpotId,
  isPassiveLocalSpotId, memorySpotLabel, memorySpotComment (plus the
  kMemorySpotIdBase / kPassiveSpotIdBase constants, used only by these)
- Client connection parsing: splitClientField, parseClientHandle,
  buildDisconnectClients ×3, cleanClientDisplayText,
  clientConnectionStatusMessage
- Misc: buildBandStackIndicatorPixmap, shortcutSequenceFromKeyEvent,
  macDaxDriverInstalled

Deliberately left behind: the shortcut-lease helpers (textInputCaptured,
shortcutGuard, leaseHolderBusy, etc.) — they read the mutable file-scope
statics s_keyboardShortcutsEnabled / s_sliderShortcutLeaseActive and are
destined for the Shortcuts TU in Phase 1. The header documents the rule:
state-coupled helpers don't belong in this file.

Two incidental notes:

- The functions lose internal linkage (static → namespace-scope external).
  No ODR risk: names are project-prefixed contextually and live in
  namespace AetherSDR.
- The long-standing Linux "-Wunused-function: macDaxDriverInstalled"
  warning at the tail of every build disappears — the function's only
  caller is inside #ifdef Q_OS_MAC, and external-linkage functions don't
  trigger that warning.

One transplant fix: the CategoryStats convenience overload of
formatNetworkSeqErrors became formatCategorySeqErrors (anon-namespace,
private to the new TU) — as an anonymous-namespace *overload* it would
have hidden the namespace-scope two-int overload from unqualified lookup.

MainWindow.cpp: 19,474 → 19,138 lines.
Verified: full build clean, all 32 ctest suites pass.

Refs #3351. Principle XI.
